dev: Modified the autotester check interface to allow for selection ancestor
This commit modifies the `AutoTester.check` method slightly by enabling
kwargs and providing a keyword argument `ancestor` which allows the
code that determines the file location of the check to be parameterized.

This commit also adds new methods `checkEval` and `throwToError` which
allows exceptions generated in the unit test to be contained and not
stop the entire unit test from completing.

This commit should NOT break any of the existing unit tests.